fix lesson number, update pupil responses

+++ SchoolSystem.Abstractions/Contracts/Commands/ScheduleCells/CreateScheduleCellCommand.cs
@@ -6,7 +6,7 @@ namespace SchoolSystem.Abstractions.Contracts.Commands.ScheduleCells
     public class CreateScheduleCellCommand : ICommand
     {
         public Guid LessonId { get; set; }
-        public byte LessonNumber { get; set; }
+        public int LessonNumber { get; set; }
         public DayOfWeek Day { get; set; }
         public int Room { get; set; }
     }
