+++ SchoolSystem.Abstractions/Contracts/Queries/ScheduleCells/ScheduleCellsResponse.cs
@@ -12,7 +12,7 @@ namespace SchoolSystem.Abstractions.Contracts.Queries.ScheduleCells
         {
             public Guid Id { get; set; }
             public Guid LessonId { get; set; }
-            public byte LessonNumber { get; set; }
+            public int LessonNumber { get; set; }
             public DayOfWeek Day { get; set; }
             public int Room { get; set; }
         }
